Update to 1.1.0 and cleanup
